+++ Facturacion/Facturacion.Application/UseCases/ProductoService.cs
@@ -1,6 +1,8 @@
 using Facturacion.Application.Dtos.Producto;
+using Facturacion.Application.Exceptions;
 using Facturacion.Application.Extensions;
 using Facturacion.Application.Interfaces;
+using Facturacion.Application.Validators.Producto;
 using Facturacion.Domain.Interfaces;
 
 namespace Facturacion.Application.UseCases
@@ -22,6 +24,8 @@ namespace Facturacion.Application.UseCases
         public async Task<ProductoDto> GetByIdAsync(int id)
         {
             var producto = await _productoRepository.GetByIdAsync(id);
+            if (producto is null)
+                throw new NotFoundException("Producto", id);
             return producto.ToDto();
         }
 
@@ -33,18 +37,36 @@ namespace Facturacion.Application.UseCases
 
         public async Task<ProductoDto> CreateAsync(CreateProductoDto productoDto)
         {
+            var validator = new CreateProductoValidatorDto();
+            var result = validator.Validate(productoDto);
+            if (!result.IsValid)
+                throw new BusinessException(string.Join(",", result.Errors.Select(e => e.ErrorMessage)));
+
             var producto = await _productoRepository.CreateAsync(productoDto.ToEntity());
             return producto.ToDto();
         }
 
         public async Task<ProductoDto> UpdateAsync(int id, UpdateProductoDto productoDto)
         {
+            var validator = new UpdateProductoValidatorDto();
+            var result = validator.Validate(productoDto);
+            if (!result.IsValid)
+                throw new BusinessException(string.Join(",", result.Errors.Select(e => e.ErrorMessage)));
+
+            var existente = await _productoRepository.GetByIdAsync(id);
+            if (existente is null)
+                throw new NotFoundException("Producto", id);
+
             var producto = await _productoRepository.UpdateAsync(id, productoDto.ToEntity());
             return producto.ToDto();
         }
 
         public async Task DeleteAsync(int id)
         {
+            var existente = await _productoRepository.GetByIdAsync(id);
+            if (existente is null)
+                throw new NotFoundException("Producto", id);
+
             await _productoRepository.DeleteAsync(id);
         }
     }
